+++ tests/plugins/test_custom_plugin.py
@@ -18,12 +18,12 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.plugins import DDPPlugin, SingleDevicePlugin
+from pytorch_lightning.plugins import DDPStrategy, SingleDevicePlugin
 from tests.helpers import BoringModel
 from tests.helpers.runif import RunIf
 
 
-class CustomParallelStrategy(DDPPlugin):
+class CustomParallelStrategy(DDPStrategy):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         # Set to None so it will be overwritten by the accelerator connector.