+++ tests/plugins/test_plugins_registry.py
@@ -16,11 +16,11 @@ import pytest
 from pytorch_lightning import Trainer
 from pytorch_lightning.plugins import (
     CheckpointIO,
-    DDPFullyShardedPlugin,
-    DDPPlugin,
+    DDPFullyShardedStrategy,
     DDPShardedPlugin,
     DDPSpawnPlugin,
     DDPSpawnShardedPlugin,
+    DDPStrategy,
     DeepSpeedPlugin,
     TPUSpawnPlugin,
     TrainingTypePluginsRegistry,
@@ -94,22 +94,22 @@ def test_tpu_spawn_debug_plugins_registry(tmpdir):
     assert isinstance(trainer.training_type_plugin, TPUSpawnPlugin)
 
 
-def test_fsdp_plugins_registry(tmpdir):
+def test_fsdp_strategys_registry(tmpdir):
 
     plugin = "fsdp"
 
     assert plugin in TrainingTypePluginsRegistry
-    assert TrainingTypePluginsRegistry[plugin]["plugin"] == DDPFullyShardedPlugin
+    assert TrainingTypePluginsRegistry[plugin]["plugin"] == DDPFullyShardedStrategy
 
     trainer = Trainer(strategy=plugin)
 
-    assert isinstance(trainer.training_type_plugin, DDPFullyShardedPlugin)
+    assert isinstance(trainer.training_type_plugin, DDPFullyShardedStrategy)
 
 
 @pytest.mark.parametrize(
     "plugin_name, plugin",
     [
-        ("ddp_find_unused_parameters_false", DDPPlugin),
+        ("ddp_find_unused_parameters_false", DDPStrategy),
         ("ddp_spawn_find_unused_parameters_false", DDPSpawnPlugin),
         ("ddp_sharded_spawn_find_unused_parameters_false", DDPSpawnShardedPlugin),
         ("ddp_sharded_find_unused_parameters_false", DDPShardedPlugin),
@@ -139,14 +139,14 @@ def test_custom_registered_training_plugin_to_strategy():
 
     custom_checkpoint_io = CustomCheckpointIO()
 
-    # Register the DDP Plugin with your custom CheckpointIO plugin
+    # Register the DDP Strategy with your custom CheckpointIO plugin
     TrainingTypePluginsRegistry.register(
         "ddp_custom_checkpoint_io",
-        DDPPlugin,
-        description="DDP Plugin with custom checkpoint io plugin",
+        DDPStrategy,
+        description="DDP Strategy with custom checkpoint io plugin",
         checkpoint_io=custom_checkpoint_io,
     )
     trainer = Trainer(strategy="ddp_custom_checkpoint_io", accelerator="cpu", devices=2)
 
-    assert isinstance(trainer.training_type_plugin, DDPPlugin)
+    assert isinstance(trainer.training_type_plugin, DDPStrategy)
     assert trainer.training_type_plugin.checkpoint_io == custom_checkpoint_io