+++ tests/trainer/test_trainer.py
@@ -40,11 +40,11 @@ from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper, Un
 from pytorch_lightning.plugins import (
     DataParallelPlugin,
     DDP2Plugin,
-    DDPFullyShardedPlugin,
-    DDPPlugin,
+    DDPFullyShardedStrategy,
     DDPShardedPlugin,
     DDPSpawnPlugin,
     DDPSpawnShardedPlugin,
+    DDPStrategy,
 )
 from pytorch_lightning.trainer.states import TrainerFn
 from pytorch_lightning.utilities import _AcceleratorType, _StrategyType
@@ -2194,11 +2194,11 @@ def test_detect_anomaly_nan(tmpdir):
             dict(_distrib_type=_StrategyType.DDP_SPAWN, _device_type=_AcceleratorType.GPU, num_gpus=2, num_processes=1),
         ),
         (
-            dict(strategy=DDPPlugin(), num_processes=2, gpus=None),
+            dict(strategy=DDPStrategy(), num_processes=2, gpus=None),
             dict(_distrib_type=_StrategyType.DDP, _device_type=_AcceleratorType.CPU, num_gpus=0, num_processes=2),
         ),
         (
-            dict(strategy=DDPPlugin(), gpus=2),
+            dict(strategy=DDPStrategy(), gpus=2),
             dict(_distrib_type=_StrategyType.DDP, _device_type=_AcceleratorType.GPU, num_gpus=2, num_processes=1),
         ),
         (
@@ -2210,7 +2210,7 @@ def test_detect_anomaly_nan(tmpdir):
             dict(_distrib_type=_StrategyType.DP, _device_type=_AcceleratorType.GPU, num_gpus=2, num_processes=1),
         ),
         (
-            dict(strategy=DDPFullyShardedPlugin(), gpus=2),
+            dict(strategy=DDPFullyShardedStrategy(), gpus=2),
             dict(
                 _distrib_type=_StrategyType.DDP_FULLY_SHARDED,
                 _device_type=_AcceleratorType.GPU,